# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticketing API",
        "description": "API for managing event tickets",
        "version": "1.0.0"
    },
    "paths": {
        "/api/check-in": {
            "post": {
                "tags": [
                    "Check-in"
                ],
                "summary": "Perform a check-in for a ticket",
                "description": "Endpoint is used to perform a check-in for a ticket by providing the ticket code.",
                "operationId": "checkIn",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "description": "Unique ticket code",
                                        "type": "string",
                                        "example": "ABC123XYZ"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful check-in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Check-in uspješan"
                                        },
                                        "ticket_code": {
                                            "type": "string",
                                            "example": "ABC123XYZ"
                                        },
                                        "checked_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-10-25T14:23:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ticket with code: ABC123XYZ not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ticket already redeemed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ticket with code: ABC123XYZ already redeemed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An unexpected error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tickets/purchase": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Purchase a ticket",
                "description": "Purchase a ticket.",
                "operationId": "purchaseTicket",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "event_id": {
                                        "description": "ID of the event",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "quantity": {
                                        "description": "Number of tickets to purchase",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket successfully purchased",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket Resource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Event": {
                "title": "Event Resource",
                "description": "An event resource associated with a ticket",
                "required": [
                    "id",
                    "name",
                    "start_time",
                    "location"
                ],
                "properties": {
                    "id": {
                        "description": "The unique ID of the event",
                        "type": "integer",
                        "example": 10
                    },
                    "name": {
                        "description": "The name of the event",
                        "type": "string",
                        "example": "Concert XYZ"
                    },
                    "start_time": {
                        "description": "The start time of the event in ISO8601 format",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-10T20:00:00Z"
                    },
                    "end_time": {
                        "description": "The end time of the event in ISO8601 format",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-10T23:00:00Z"
                    },
                    "location": {
                        "description": "The location where the event is being held",
                        "type": "string",
                        "example": "Madison Square Garden"
                    }
                },
                "type": "object"
            },
            "Ticket Resource": {
                "title": "Ticket Resource",
                "description": "A single ticket resource",
                "required": [
                    "id",
                    "category",
                    "price",
                    "sold"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the Ticket",
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "description": "The category of the ticket",
                        "type": "string",
                        "example": "Concert A"
                    },
                    "price": {
                        "description": "Price of the ticket",
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "sold": {
                        "description": "Number of tickets sold",
                        "type": "integer",
                        "example": 3
                    },
                    "code": {
                        "description": "Unique ticket code",
                        "type": "string",
                        "example": "ABC123XYZ"
                    },
                    "event_id": {
                        "description": "Event's id",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "User's id",
                        "type": "integer",
                        "example": 1
                    },
                    "user_name": {
                        "description": "User's name",
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "purchase_date": {
                        "description": "Time of purchase",
                        "type": "date-time",
                        "example": "2025-11-22T10:14:51.000000Z"
                    },
                    "event": {
                        "$ref": "#/components/schemas/Event"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Check-in",
            "description": "Check-in"
        },
        {
            "name": "Tickets",
            "description": "Tickets"
        }
    ]
}